requirements.txt
python-telegram-bot==20.7
apscheduler==3.11.1
aiohttp
requests
beautifulsoup4
fastapi
uvicorn
jinja2
